services/data_collector/requirements-data-collector.txt
python-dotenv==1.0.1
requests==2.32.3
supabase==2.12.0
web3==7.7.0
beautifulsoup4==4.12.2
